Implemented support for multiple subscriptions

+++ snapcleanup/cli.py
@@ -8,8 +8,7 @@ cli = typer.Typer(help="Azure Disk Snapshot Cleanup", no_args_is_help=True)
 snap_cleanup = DiskSnapshotCleanup(
     client_id=settings.AZURE_CLIENT_ID,
     client_secret=settings.AZURE_CLIENT_SECRET,
-    tenant_id=settings.AZURE_TENANT_ID,
-    subscription_id=settings.AZURE_SUBSCRIPTION_ID,
+    tenant_id=settings.AZURE_TENANT_ID
 )
 
 
@@ -23,7 +22,7 @@ def list_subscriptions(
     ),
     verbose: Optional[bool] = typer.Option(
         False, "--verbose", "-v", help="Show more logs"
-    ),
+    )
 ):
     """List all subscriptions that user has access."""
     subscriptions = snap_cleanup.list_subscriptions()
@@ -31,7 +30,7 @@ def list_subscriptions(
     snap_cleanup.print_table(
         table_title="Subscription List",
         headers=fields,
-        object_list=subscriptions,
+        object_list=subscriptions
     )
     if isinstance(csv, str):
         snap_cleanup.create_csv_file(
@@ -39,7 +38,7 @@ def list_subscriptions(
             headers=fields,
             object_list=subscriptions,
             delimiter=";",
-            verbose=verbose,
+            verbose=verbose
         )
     if isinstance(json, str):
         snap_cleanup.create_json_file(
@@ -47,7 +46,7 @@ def list_subscriptions(
             headers=fields,
             object_list=subscriptions,
             indent_spaces=4,
-            verbose=verbose,
+            verbose=verbose
         )
 
 
@@ -61,32 +60,35 @@ def list_resource_groups(
     ),
     verbose: Optional[bool] = typer.Option(
         False, "--verbose", "-v", help="Show more logs"
-    ),
+    )
 ):
     """List all resource groups."""
-    resource_groups = snap_cleanup.list_resource_groups()
-    fields = ["name", "location"]
-    snap_cleanup.print_table(
-        table_title="Resource Group List",
-        headers=fields,
-        object_list=resource_groups,
-    )
-    if isinstance(csv, str):
-        snap_cleanup.create_csv_file(
-            file_path=csv,
-            headers=fields,
-            object_list=resource_groups,
-            delimiter=";",
-            verbose=verbose,
-        )
-    if isinstance(json, str):
-        snap_cleanup.create_json_file(
-            file_path=json,
-            headers=fields,
-            object_list=resource_groups,
-            indent_spaces=4,
-            verbose=verbose,
-        )
+    for subscription_id in settings.AZURE_SUBSCRIPTION_ID:
+        subscription_info = snap_cleanup.get_subscription(subscription_id)
+        if snap_cleanup.set_subscription(subscription_id):
+            resource_groups = snap_cleanup.list_resource_groups()
+            fields = ["name", "location"]
+            snap_cleanup.print_table(
+                table_title=f"Resource Group List: {subscription_info.name}",
+                headers=fields,
+                object_list=resource_groups,
+            )
+            if isinstance(csv, str):
+                snap_cleanup.create_csv_file(
+                    file_path=csv,
+                    headers=fields,
+                    object_list=resource_groups,
+                    delimiter=";",
+                    verbose=verbose
+                )
+            if isinstance(json, str):
+                snap_cleanup.create_json_file(
+                    file_path=json,
+                    headers=fields,
+                    object_list=resource_groups,
+                    indent_spaces=4,
+                    verbose=verbose
+                )
 
 
 @cli.command("list-snap")
@@ -99,38 +101,41 @@ def list_snapshots(
     ),
     verbose: Optional[bool] = typer.Option(
         False, "--verbose", "-v", help="Show more logs"
-    ),
+    )
 ):
     """List all snapshots."""
-    snapshots = snap_cleanup.list_snapshots()
-    fields = [
-        "resource_group",
-        "name",
-        "location",
-        "created_date",
-        "ttl_tag_value",
-    ]
-    snap_cleanup.print_table(
-        table_title="Snapshot List",
-        headers=fields,
-        object_list=snapshots,
-    )
-    if isinstance(csv, str):
-        snap_cleanup.create_csv_file(
-            file_path=csv,
-            headers=fields,
-            object_list=snapshots,
-            delimiter=";",
-            verbose=verbose,
-        )
-    if isinstance(json, str):
-        snap_cleanup.create_json_file(
-            file_path=json,
-            headers=fields,
-            object_list=snapshots,
-            indent_spaces=4,
-            verbose=verbose,
-        )
+    for subscription_id in settings.AZURE_SUBSCRIPTION_ID:
+        subscription_info = snap_cleanup.get_subscription(subscription_id)
+        if snap_cleanup.set_subscription(subscription_id):
+            snapshots = snap_cleanup.list_snapshots()
+            fields = [
+                "resource_group",
+                "name",
+                "location",
+                "created_date",
+                "ttl_tag_value",
+            ]
+            snap_cleanup.print_table(
+                table_title=f"Snapshot List: {subscription_info.name}",
+                headers=fields,
+                object_list=snapshots,
+            )
+            if isinstance(csv, str):
+                snap_cleanup.create_csv_file(
+                    file_path=csv,
+                    headers=fields,
+                    object_list=snapshots,
+                    delimiter=";",
+                    verbose=verbose
+                )
+            if isinstance(json, str):
+                snap_cleanup.create_json_file(
+                    file_path=json,
+                    headers=fields,
+                    object_list=snapshots,
+                    indent_spaces=4,
+                    verbose=verbose
+                )
 
 
 @cli.command("update-snap-tag")
@@ -155,42 +160,45 @@ def update_snap_tag(
     ),
 ):
     """Add time-to-live tag on snapshots that do not already have."""
-    snapshots = snap_cleanup.list_snapshots()
-    updated_snapshots = snap_cleanup.update_snapshot_tag(
-        list_snapshots=snapshots,
-        ttl_tag_name=ttl_tag_name,
-        ttl_days=ttl_days,
-        dry_run=dry_run,
-    )
-    fields = [
-        "resource_group",
-        "name",
-        "location",
-        "created_date",
-        "ttl_tag_value",
-        "action",
-    ]
-    snap_cleanup.print_table(
-        table_title="Updated Snapshots",
-        headers=fields,
-        object_list=updated_snapshots,
-    )
-    if isinstance(csv, str):
-        snap_cleanup.create_csv_file(
-            file_path=csv,
-            headers=fields,
-            object_list=updated_snapshots,
-            delimiter=";",
-            verbose=verbose,
-        )
-    if isinstance(json, str):
-        snap_cleanup.create_json_file(
-            file_path=json,
-            headers=fields,
-            object_list=updated_snapshots,
-            indent_spaces=4,
-            verbose=verbose,
-        )
+    for subscription_id in settings.AZURE_SUBSCRIPTION_ID:
+        subscription_info = snap_cleanup.get_subscription(subscription_id)
+        if snap_cleanup.set_subscription(subscription_id):
+            snapshots = snap_cleanup.list_snapshots()
+            updated_snapshots = snap_cleanup.update_snapshot_tag(
+                list_snapshots=snapshots,
+                ttl_tag_name=ttl_tag_name,
+                ttl_days=ttl_days,
+                dry_run=dry_run,
+            )
+            fields = [
+                "resource_group",
+                "name",
+                "location",
+                "created_date",
+                "ttl_tag_value",
+                "action",
+            ]
+            snap_cleanup.print_table(
+                table_title=f"Updated Snapshots: {subscription_info.name}",
+                headers=fields,
+                object_list=updated_snapshots,
+            )
+            if isinstance(csv, str):
+                snap_cleanup.create_csv_file(
+                    file_path=csv,
+                    headers=fields,
+                    object_list=updated_snapshots,
+                    delimiter=";",
+                    verbose=verbose
+                )
+            if isinstance(json, str):
+                snap_cleanup.create_json_file(
+                    file_path=json,
+                    headers=fields,
+                    object_list=updated_snapshots,
+                    indent_spaces=4,
+                    verbose=verbose
+                )
 
 
 @cli.command("delete-snap")
@@ -209,36 +217,39 @@ def delete_snap_tag(
     ),
 ):
     """Delete all Add time-to-live tag on snapshots that do not already have."""
-    snapshots = snap_cleanup.list_snapshots()
-    deleted_snapshots = snap_cleanup.delete_snapshots(
-        list_snapshots=snapshots, dry_run=dry_run
-    )
-    fields = [
-        "resource_group",
-        "name",
-        "location",
-        "created_date",
-        "ttl_tag_value",
-        "action",
-    ]
-    snap_cleanup.print_table(
-        table_title="Deleted Snapshots",
-        headers=fields,
-        object_list=deleted_snapshots,
-    )
-    if isinstance(csv, str):
-        snap_cleanup.create_csv_file(
-            file_path=csv,
-            headers=fields,
-            object_list=deleted_snapshots,
-            delimiter=";",
-            verbose=verbose,
-        )
-    if isinstance(json, str):
-        snap_cleanup.create_json_file(
-            file_path=json,
-            headers=fields,
-            object_list=deleted_snapshots,
-            indent_spaces=4,
-            verbose=verbose,
-        )
+    for subscription_id in settings.AZURE_SUBSCRIPTION_ID:
+        subscription_info = snap_cleanup.get_subscription(subscription_id)
+        if snap_cleanup.set_subscription(subscription_id):
+            snapshots = snap_cleanup.list_snapshots()
+            deleted_snapshots = snap_cleanup.delete_snapshots(
+                list_snapshots=snapshots, dry_run=dry_run
+            )
+            fields = [
+                "resource_group",
+                "name",
+                "location",
+                "created_date",
+                "ttl_tag_value",
+                "action",
+            ]
+            snap_cleanup.print_table(
+                table_title=f"Deleted Snapshots: {subscription_info.name}",
+                headers=fields,
+                object_list=deleted_snapshots,
+            )
+            if isinstance(csv, str):
+                snap_cleanup.create_csv_file(
+                    file_path=csv,
+                    headers=fields,
+                    object_list=deleted_snapshots,
+                    delimiter=";",
+                    verbose=verbose
+                )
+            if isinstance(json, str):
+                snap_cleanup.create_json_file(
+                    file_path=json,
+                    headers=fields,
+                    object_list=deleted_snapshots,
+                    indent_spaces=4,
+                    verbose=verbose
+                )
